committing work to restore aoc
