--- 2019/problem1/answer1.js
@@ -1,9 +1,0 @@
-var answer=input.split("\n");
-var sum =0;
-answer.forEach(calcFunc);
-
-function calcFunc(item){
-sum+= ((Math.floor(item/3))-2)
-}
-
-document.getElementById("output").innerHTML = sum
