+++ answers/2018/problem1/answer1.2.js
