+++ answers/2018/problem1/input1.txt
