+++ answers/2018/problem2/answer2.js
