+++ answers/2018/problem2/input2.txt
