+++ answers/2018/problem3/answer3.js
