+++ answers/2018/problem3/input3.txt
