+++ answers/2018/problem6/2018.problem6.answer1.txt
