--- index.html
@@ -1,50 +1,0 @@
-<meta content="text/html;charset=utf-8" http-equiv="Content-Type">
-<meta content="utf-8" http-equiv="encoding">
-<!DOCTYPE html>
-<html>
-    <body bgcolor="90ee90">
-        <div style="text-align: center;">
-        <h1>Advent of Code</h1>
-        <pre>
-<font color="#d4af37">
-/\
-<  >
-\/</font><font color="darkgreen"></font>
-/\
-/  \
-/<font color="white">++++</font>\
-/  <font color="red">()</font>  \
-/      \
-/<font color="red">~`~`~`~`</font>\
-/  <font color="white">()</font>  <font color="red">()</font>  \
-/          \
-/<font color="white">*&*&*&*&*&*&</font>\
-/  <font color="red">()  <font color="white">()</font>  ()</font>  \
-/              \
-/<font color="red">++++++++++++++++</font>\
-/  <font color="white">()  <font color="red">()</font>  ()  <font color="red">()</font></font>  \
-/                  \
-/<font color="white">~`~`~`~`~`~`~`~`~`~`</font>\
-/  <font color="red">()  <font color="white">()</font>  ()  <font color="white">()</font>  ()</font>  \
-/                      \
-/ <font color="red">*&*&*&*&*&*&*&*&*&*&*&</font> \
-/,.,.,.,.,.,.,.,.,.,.,.,.,.\
-</font><font color="brown">|   |</font>
-<font color="darkred">|`````|
-\_____/</font>
-        </pre>
-        <br><br>
-
-        <input type='file' id='fileinput' accept='text/plain' onchange='setInfo(event)'><br><br>
-<button id='visualizebtn' onclick='vis2D()' style="display:none">Visualize</button>        <br>
-        <p>Output: </p><p id="output"></p>
-</div>
-<script src="code.js">
-</script>
-
-
-<script id="script"></script>
-<script src="visualize.js"></script>
-</body>
-</html> 
-
