+++ tools/makefolder.ps1
@@ -1,3 +1,3 @@
 for($i = 1; $i -lt 26; $i++){ 
-New-Item -Name "problem$i" -ItemType "directory"
+New-Item -Name "day$i" -ItemType "directory"
 }
